Fix #1972 run template.fixup via pkcli.srw Fixups not requiring template moved from srw_fixup.py to sim_data/srw.py Fixups requiring template called from template in a subprocess remove "template_fixup"
