+++ sirepo/sim_data/srw.py
@@ -148,9 +148,73 @@ class SimData(sirepo.sim_data.SimDataBase):
             for f in 'horizontalPosition', 'verticalPosition', 'driftCalculationMethod', 'drift':
                 if f in e:
                     del e[f]
+        cls.__fixup_old_data_beamline(data)
+        if 'horizontalDeflectingParameter' not in dm.undulator:
+            cls.__fixup_old_data_by_template(data)
+            dm = data.models
+        hv = ('horizontalPosition', 'horizontalRange', 'verticalPosition', 'verticalRange')
+        if 'samplingMethod' not in dm.simulation:
+            simulation = dm.simulation
+            simulation.samplingMethod = 1 if simulation.sampleFactor > 0 else 2
+            for k in hv:
+                simulation[k] = dm.initialIntensityReport[k]
+        if 'horizontalPosition' in dm.initialIntensityReport:
+            for k in dm:
+                if k == 'sourceIntensityReport' or k == 'initialIntensityReport' or cls.is_watchpoint(k):
+                    for f in hv:
+                        del dm[k][f]
+        if 'indexFile' in data.models.tabulatedUndulator:
+            del data.models.tabulatedUndulator['indexFile']
+        u = dm.undulator
+        if 'effectiveDeflectingParameter' not in u and 'horizontalDeflectingParameter' in u:
+            u.effectiveDeflectingParameter = math.sqrt(
+                u.horizontalDeflectingParameter ** 2 + u.verticalDeflectingParameter ** 2,
+            )
+        for k in (
+            'photonEnergy',
+            'horizontalPointCount',
+            'horizontalPosition',
+            'horizontalRange',
+            'sampleFactor',
+            'samplingMethod',
+            'verticalPointCount',
+            'verticalPosition',
+            'verticalRange',
+        ):
+            if k not in dm.sourceIntensityReport:
+                dm.sourceIntensityReport[k] = dm.simulation[k]
+        if 'photonEnergy' not in dm.gaussianBeam:
+            dm.gaussianBeam.photonEnergy = dm.simulation.photonEnergy
+        if 'longitudinalPosition' in dm.tabulatedUndulator:
+            u = dm.tabulatedUndulator
+            for k in (
+                'undulatorParameter',
+                'period',
+                'longitudinalPosition',
+                'horizontalAmplitude',
+                'horizontalSymmetry',
+                'horizontalInitialPhase',
+                'verticalAmplitude',
+                'verticalSymmetry',
+                'verticalInitialPhase',
+            ):
+                if k in u:
+                    if cls.srw_is_tabulated_undulator_source(dm.simulation):
+                        dm.undulator[k] = u[k]
+                    del u[k]
+        if 'name' not in dm['tabulatedUndulator']:
+            u = dm.tabulatedUndulator
+            u.name = u.undulatorSelector = 'Undulator'
+        if dm.tabulatedUndulator.get('id', '1') == '1':
+            dm.tabulatedUndulator.id = '{} 1'.format(dm.simulation.simulationId)
+        if len(dm.postPropagation) == 9:
+            dm.postPropagation += [0, 0, 0, 0, 0, 0, 0, 0]
+            for i in dm.propagation:
+                for r in dm.propagation[i]:
+                    r += [0, 0, 0, 0, 0, 0, 0, 0]
+        if 'electronBeams' in dm:
+            del dm['electronBeams']
         cls._organize_example(data)
-        cls._template_fixup_set(data)
-
 
     @classmethod
     def lib_file_name_with_type(cls, filename, file_type):
@@ -181,21 +245,25 @@ class SimData(sirepo.sim_data.SimDataBase):
             dm.simulation.folder = '/'
 
     @classmethod
-    def srw_lib_file_paths_for_type(cls, file_type, op, want_user_lib_dir):
-        """Search for files of type"""
-        from sirepo import simulation_db
-
-        res = []
-        for e in cls.SRW_FILE_TYPE_EXTENSIONS[file_type]:
-            for f in cls._lib_file_list('*.{}'.format(e), want_user_lib_dir=want_user_lib_dir):
-                x = op(f)
-                if x:
-                    res.append(x)
-        return res
-
-    @classmethod
     def srw_compute_crystal_grazing_angle(cls, model):
         model.grazingAngle = math.acos(math.sqrt(1 - model.tvx ** 2 - model.tvy ** 2)) * 1e3
+
+    @classmethod
+    def srw_find_closest_angle(cls, angle, allowed_angles):
+        """Find closest string value from the input list to
+           the specified angle (in radians).
+        """
+        def _wrap(a):
+            """Convert an angle to constraint it between -pi and pi.
+               See https://stackoverflow.com/a/29237626/4143531 for details.
+            """
+            return numpy.arctan2(numpy.sin(a), numpy.cos(a))
+
+        angles = numpy.array([float(x) for x in allowed_angles])
+        threshold = numpy.min(numpy.diff(angles))
+        return allowed_angles[
+            numpy.where(numpy.abs(_wrap(angle) - angles) < threshold / 2.0)[0][0]
+        ]
 
     @classmethod
     def srw_format_float(cls, v):
@@ -264,6 +332,17 @@ class SimData(sirepo.sim_data.SimDataBase):
     @classmethod
     def srw_is_valid_file_type(cls, file_type, path):
         return path.ext[1:] in cls.SRW_FILE_TYPE_EXTENSIONS.get(file_type, tuple())
+
+    @classmethod
+    def srw_lib_file_paths_for_type(cls, file_type, op, want_user_lib_dir):
+        """Search for files of type"""
+        res = []
+        for e in cls.SRW_FILE_TYPE_EXTENSIONS[file_type]:
+            for f in cls._lib_file_list('*.{}'.format(e), want_user_lib_dir=want_user_lib_dir):
+                x = op(f)
+                if x:
+                    res.append(x)
+        return res
 
     @classmethod
     def srw_predefined(cls):
@@ -360,3 +439,109 @@ class SimData(sirepo.sim_data.SimDataBase):
                     if m[k] and t in ('MirrorFile', 'ImageFile'):
                         res.append(m[k])
         return res
+
+    @classmethod
+    def __fixup_old_data_by_template(cls, data):
+        import pykern.pkcompat
+        import pykern.pkjson
+        import sirepo.simulation_db
+        import subprocess
+        import os
+        import sys
+
+        with sirepo.simulation_db.tmp_dir() as d:
+            f = d.join('in.json')
+            pykern.pkjson.dump_pretty(data, filename=f, pretty=False)
+            try:
+                #TODO(robnagler) find a better way to do this
+                e = PKDict(os.environ).pkupdate(
+                    SIREPO_AUTH_LOGGED_IN_USER=str(sirepo.auth.logged_in_user()),
+                    SIREPO_SRDB_ROOT=str(sirepo.srdb.root()),
+                )
+                n = subprocess.check_output(
+                    ['sirepo', 'srw', 'fixup_old_data', str(f)],
+                    stderr=subprocess.STDOUT,
+                    env=e,
+                )
+            except subprocess.CalledProcessError as e:
+                pkdlog('sirepo.pkcli.srw.fixup_old_data failed: {}', e.output)
+                raise
+            data.clear()
+            try:
+                data.update(pykern.pkjson.load_any(n))
+            except Exception as e:
+                pkdlog('unable to parse json={}', n)
+                raise
+
+    @classmethod
+    def __fixup_old_data_beamline(cls, data):
+        dm = data.models
+        for i in dm.beamline:
+            t = i.type
+            if t == 'ellipsoidMirror':
+                if 'firstFocusLength' not in i:
+                    i.firstFocusLength = i.position
+            if t in ('grating', 'ellipsoidMirror', 'sphericalMirror', 'toroidalMirror'):
+                if 'grazingAngle' not in i:
+                    angle = 0
+                    if i.normalVectorX:
+                        angle = math.acos(abs(float(i.normalVectorX))) * 1000
+                    elif i.normalVectorY:
+                        angle = math.acos(abs(float(i.normalVectorY))) * 1000
+                    i.grazingAngle = angle
+            if 'grazingAngle' in i and 'normalVectorX' in i and 'autocomputeVectors' not in i:
+                i.autocomputeVectors = '1'
+            if t == 'crl':
+                for k, v in PKDict(
+                    material='User-defined',
+                    method='server',
+                    absoluteFocusPosition=None,
+                    focalDistance=None,
+                    tipRadius=float(i.radius) * 1e6,  # m -> um
+                    tipWallThickness=float(i.wallThickness) * 1e6,  # m -> um
+                ).items():
+                    if k not in i:
+                        i[k] = v
+            if t == 'crystal':
+                if 'diffractionAngle' not in i:
+                    allowed_angles = [x[0] for x in cls.schema().enum.DiffractionPlaneAngle]
+                    i.diffractionAngle = cls.srw_find_closest_angle(i.grazingAngle or 0, allowed_angles)
+                    if i.tvx == '':
+                        i.tvx = i.tvy = 0
+                    cls.srw_compute_crystal_grazing_angle(i)
+            if t == 'sample':
+                if 'horizontalCenterCoordinate' not in i:
+                    i.horizontalCenterCoordinate = cls.schema().model.sample.horizontalCenterCoordinate[2]
+                    i.verticalCenterCoordinate = cls.schema().model.sample.verticalCenterCoordinate[2]
+                if 'cropArea' not in i:
+                    for f in (
+                        'areaXEnd',
+                        'areaXStart',
+                        'areaYEnd',
+                        'areaYStart',
+                        'backgroundColor',
+                        'cropArea',
+                        'cutoffBackgroundNoise',
+                        'invert',
+                        'outputImageFormat',
+                        'rotateAngle',
+                        'rotateReshape',
+                        'shiftX',
+                        'shiftY',
+                        'tileColumns',
+                        'tileImage',
+                        'tileRows',
+                    ):
+                        i[f] = cls.schema().model.sample[f][2]
+                if 'transmissionImage' not in i:
+                    i.transmissionImage = cls.schema().model.sample.transmissionImage[2]
+            if t in ('crl', 'grating', 'ellipsoidMirror', 'sphericalMirror') \
+                and 'horizontalOffset' not in i:
+                i.horizontalOffset = 0
+                i.verticalOffset = 0
+            if 'autocomputeVectors' in i:
+                if i.autocomputeVectors == '0':
+                    i.autocomputeVectors = 'none'
+                elif i.autocomputeVectors == '1':
+                    i.autocomputeVectors = 'vertical' if i.normalVectorX == 0 else 'horizontal'
+            cls.update_model_defaults(i, t)
